remove torch model defination to seperate file
