+++ neural_compressor/adaptor/torch_utils/smooth_quant.py
@@ -164,74 +164,6 @@ def set_module(model, key, new_module):
     setattr(module, attrs[-1], new_module)
 
 
-class QDQLinear(torch.nn.Module):
-    def __init__(self, module, scale, zero_point, dtype):
-        super().__init__()
-        import torch.ao.nn.quantized as nnq
-        self.add_module('quant', nnq.Quantize(scale, zero_point, dtype))
-        self.add_module('dequant', nnq.DeQuantize())
-        self.add_module('module', module)
-        self.qdq_weight()
-
-    def forward(self, X):
-        X = self.quant(X)
-        X = self.dequant(X)
-        X = self.module(X)
-        return X
-
-    def qdq_weight(self):
-        # update weight w/ QDQ
-        weith_qdq = quant_dequant_w(self.module)
-        self.module.weight = torch.nn.Parameter(weith_qdq)
-
-
-class SQLinearWrapper(torch.nn.Module):
-    def __init__(self, module, input_scale, input_minmax, dtype=torch.quint8):
-        super().__init__()
-        self.input_scale = input_scale
-        self.dtype = dtype
-        # calculate and only save scale, zero_point to avoid memory usage
-        self.scale, self.zero_point = self._calculate_qparams(input_scale, input_minmax, dtype)
-        self.add_module('sq_linear', module)
-        self.ipex = False  # a flag used for ipex inference
-
-    def forward(self, X):
-        if self.ipex:
-            X = self.sq_linear(X)
-        else:
-            X = torch.mul(X, self.input_scale)
-            X = self.sq_linear(X)
-        return X
-
-    def _calculate_qparams(self, input_scale, input_minmax, dtype=torch.quint8):
-        # calculate scale and zero_point
-        if dtype == torch.quint8:
-            quant_min, quant_max = 0, 255
-        min_val = torch.min(input_minmax[0] * input_scale)
-        max_val = torch.max(input_minmax[1] * input_scale)
-        # work when min_val bigger than zero.
-        min_val_neg = torch.min(min_val, torch.zeros_like(min_val))
-        max_val_pos = torch.max(max_val, torch.zeros_like(max_val))
-        scale = (max_val_pos - min_val_neg) / float(quant_max - quant_min)
-        scale = torch.max(scale, torch.tensor([torch.finfo(torch.float32).eps]))
-        zero_point = quant_min - torch.round(min_val_neg / scale).to(torch.int)
-        zero_point = torch.clamp(zero_point, quant_min, quant_max)
-        return scale, zero_point
-
-    def _get_weight_scale(self):
-        # get weight scale and zero_point
-        from torch.ao.quantization.observer import default_per_channel_weight_observer
-        obs = default_per_channel_weight_observer()
-        obs(self.sq_linear.weight)
-        scale, _ = obs.calculate_qparams()
-        return scale
-
-    def _recover_sq_linear(self):
-        # remove mul and reset sq_linear for ipex inference
-        scale = self.input_scale.view(1, self.input_scale.shape[0])
-        with torch.no_grad():
-            self.sq_linear.weight *= scale
-
 class TorchSmoothQuant:
     """
     Fake input channel quantization, for more details please refer to
@@ -460,6 +392,7 @@ class TorchSmoothQuant:
         :return:
         """
         layer = get_module(self.model, layer_name)
+        from .model_wrapper import SQLinearWrapper
         if isinstance(layer, SQLinearWrapper):
             layer = layer.sq_linear
         scale = self._reshape_scale_for_weight(layer, scale)
@@ -473,6 +406,7 @@ class TorchSmoothQuant:
         :param scale: The scale to be absorbed
         :return:
         """
+        from .model_wrapper import SQLinearWrapper
         layer = get_module(self.model, layer_name)
         if self.insert_mul:
             if isinstance(layer, SQLinearWrapper):
@@ -634,6 +568,7 @@ class TorchSmoothQuant:
         """
         logger.info("auto tuning alpha")
         import copy
+        from .model_wrapper import SQLinearWrapper
         alpha_scale = 100
         alpha_space = list(range(round(alpha_min * alpha_scale), round((alpha_max + alpha_step) * alpha_scale),
                                  round(alpha_step * alpha_scale)))
